Fold `TokenValidator` into `AuthenticationContextFactory`
