+++ src/Tgstation.Server.Host/Security/AuthenticationContextFactory.cs
@@ -1,20 +1,27 @@
 ﻿using System;
+using System.Globalization;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using Microsoft.IdentityModel.JsonWebTokens;
+using Microsoft.IdentityModel.Tokens;
 
+using Tgstation.Server.Api;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.Database;
 using Tgstation.Server.Host.Models;
+using Tgstation.Server.Host.Utils;
 
 namespace Tgstation.Server.Host.Security
 {
 	/// <inheritdoc cref="IAuthenticationContext" />
-	sealed class AuthenticationContextFactory : IAuthenticationContextFactory, IDisposable
+	sealed class AuthenticationContextFactory : ITokenValidator, IDisposable
 	{
 		/// <summary>
 		/// The <see cref="IAuthenticationContext"/> the <see cref="AuthenticationContextFactory"/> created.
@@ -47,6 +54,11 @@ namespace Tgstation.Server.Host.Security
 		readonly AuthenticationContext currentAuthenticationContext;
 
 		/// <summary>
+		/// The <see cref="ApiHeaders"/> for the <see cref="AuthenticationContextFactory"/>.
+		/// </summary>
+		readonly ApiHeaders? apiHeaders;
+
+		/// <summary>
 		/// 1 if <see cref="currentAuthenticationContext"/> was initialized, 0 otherwise.
 		/// </summary>
 		int initialized;
@@ -56,16 +68,21 @@ namespace Tgstation.Server.Host.Security
 		/// </summary>
 		/// <param name="databaseContext">The value of <see cref="databaseContext"/>.</param>
 		/// <param name="identityCache">The value of <see cref="identityCache"/>.</param>
+		/// <param name="apiHeadersProvider">The <see cref="IApiHeadersProvider"/> containing the value of <see cref="apiHeaders"/>.</param>
 		/// <param name="swarmConfigurationOptions">The <see cref="IOptions{TOptions}"/> containing the value of <see cref="swarmConfiguration"/>.</param>
 		/// <param name="logger">The value of <see cref="logger"/>.</param>
 		public AuthenticationContextFactory(
 			IDatabaseContext databaseContext,
 			IIdentityCache identityCache,
+			IApiHeadersProvider apiHeadersProvider,
 			IOptions<SwarmConfiguration> swarmConfigurationOptions,
 			ILogger<AuthenticationContextFactory> logger)
 		{
 			this.databaseContext = databaseContext ?? throw new ArgumentNullException(nameof(databaseContext));
 			this.identityCache = identityCache ?? throw new ArgumentNullException(nameof(identityCache));
+			ArgumentNullException.ThrowIfNull(apiHeadersProvider);
+
+			apiHeaders = apiHeadersProvider.ApiHeaders;
 			swarmConfiguration = swarmConfigurationOptions?.Value ?? throw new ArgumentNullException(nameof(swarmConfigurationOptions));
 			this.logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
@@ -76,10 +93,47 @@ namespace Tgstation.Server.Host.Security
 		public void Dispose() => currentAuthenticationContext.Dispose();
 
 		/// <inheritdoc />
-		public async ValueTask<IAuthenticationContext> CreateAuthenticationContext(long userId, long? instanceId, DateTimeOffset notBefore, CancellationToken cancellationToken)
+		public async Task ValidateToken(TokenValidatedContext tokenValidatedContext, CancellationToken cancellationToken)
 		{
+			ArgumentNullException.ThrowIfNull(tokenValidatedContext);
+
+			if (tokenValidatedContext.SecurityToken is not JsonWebToken jwt)
+				throw new ArgumentException($"Expected {nameof(tokenValidatedContext)} to contain a {nameof(JsonWebToken)}!", nameof(tokenValidatedContext));
+
 			if (Interlocked.Exchange(ref initialized, 1) != 0)
 				throw new InvalidOperationException("Authentication context has already been loaded");
+
+			var principal = new ClaimsPrincipal(new ClaimsIdentity(jwt.Claims));
+
+			var userIdClaim = principal.FindFirst(JwtRegisteredClaimNames.Sub);
+			if (userIdClaim == default)
+				throw new InvalidOperationException($"Missing '{JwtRegisteredClaimNames.Sub}' claim!");
+
+			long userId;
+			try
+			{
+				userId = Int64.Parse(userIdClaim.Value, CultureInfo.InvariantCulture);
+			}
+			catch (Exception e)
+			{
+				throw new InvalidOperationException("Failed to parse user ID!", e);
+			}
+
+			var nbfClaim = principal.FindFirst(JwtRegisteredClaimNames.Nbf);
+			if (nbfClaim == default)
+				throw new InvalidOperationException($"Missing '{JwtRegisteredClaimNames.Nbf}' claim!");
+
+			DateTimeOffset notBefore;
+			try
+			{
+				notBefore = new DateTimeOffset(
+					EpochTime.DateTime(
+						Int64.Parse(nbfClaim.Value, CultureInfo.InvariantCulture)));
+			}
+			catch (Exception ex)
+			{
+				throw new InvalidOperationException("Failed to parse nbf!", ex);
+			}
 
 			var user = await databaseContext
 				.Users
@@ -93,8 +147,8 @@ namespace Tgstation.Server.Host.Security
 				.FirstOrDefaultAsync(cancellationToken);
 			if (user == default)
 			{
-				logger.LogWarning("Unable to find user with ID {userId}!", userId);
-				return currentAuthenticationContext;
+				tokenValidatedContext.Fail($"Unable to find user with ID {userId}!");
+				return;
 			}
 
 			ISystemIdentity? systemIdentity;
@@ -104,8 +158,8 @@ namespace Tgstation.Server.Host.Security
 			{
 				if (user.LastPasswordUpdate.HasValue && user.LastPasswordUpdate >= notBefore)
 				{
-					logger.LogDebug("Rejecting token for user {userId} created before last password update: {lastPasswordUpdate}", userId, user.LastPasswordUpdate.Value);
-					return currentAuthenticationContext;
+					tokenValidatedContext.Fail($"Rejecting token for user {userId} created before last modification: {user.LastPasswordUpdate.Value}");
+					return;
 				}
 
 				systemIdentity = null;
@@ -115,6 +169,7 @@ namespace Tgstation.Server.Host.Security
 			try
 			{
 				InstancePermissionSet? instancePermissionSet = null;
+				var instanceId = apiHeaders?.InstanceId;
 				if (instanceId.HasValue)
 				{
 					instancePermissionSet = await databaseContext.InstancePermissionSets
@@ -131,7 +186,6 @@ namespace Tgstation.Server.Host.Security
 					systemIdentity,
 					user,
 					instancePermissionSet);
-				return currentAuthenticationContext;
 			}
 			catch
 			{
